requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
sqlalchemy==2.0.23
python-dotenv==1.0.0

